requirements.txt
pyjwt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django
django-cors-headers
fonttools==4.50.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
shiboken6==6.6.3
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
